peek_ret: Get ARM working
ARM can return a few ways:
* Popping a bunch of registers off the stack
* Branching to LR

We can detect return in the former case if PC is the last arg, in which
case to get the return address we have to use the number of operands to
see how far into the stack it is found.

If it's just branching to LR, we just return the value from the
register.